solver-cuestionarios/requirements.txt
google-genai
mss
Pillow
python-dotenv
imagehash
fpdf2